@@ -1,16 +1,16 @@
 module github.com/defektive/requrse
 
-go 1.23.4
+go 1.24
 
 require (
 	github.com/gorilla/websocket v1.5.3
 	github.com/itchyny/gojq v0.12.17
-	github.com/spf13/cobra v1.9.1
+	github.com/spf13/cobra v1.10.1
 	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/itchyny/timefmt-go v0.1.6 // indirect
-	github.com/spf13/pflag v1.0.6 // indirect
+	github.com/itchyny/timefmt-go v0.1.7 // indirect
+	github.com/spf13/pflag v1.0.10 // indirect
 )
